+++ src/utils.rs
@@ -1,8 +1,9 @@
 pub(crate) use std::path::Path;
 
+use anyhow::{anyhow, Context};
 pub(crate) use clap::crate_version;
 
-use crate::paths;
+use crate::configuration::manifest::BaseGameVersion;
 
 /// Returns `true` when the specified directory is empty.
 /// Returns an [Err] if
@@ -14,10 +15,25 @@ pub(crate) fn empty_dir(path: &Path) -> std::io::Result<bool> {
     Ok(path.read_dir()?.next().is_none())
 }
 
-pub(crate) fn version_as_array(version: &str) -> (usize, usize, usize) {
-    let parts: Vec<usize> = version
-        .split('.')
-        .map(|x| x.parse().expect("invalid version format"))
-        .collect();
-    (parts[0], parts[1], parts[2])
+pub(crate) fn version_as_array(version: &str) -> anyhow::Result<(usize, usize, usize)> {
+    let mut parsed_parts: Vec<usize> = Vec::new();
+    let parts: Vec<&str> = version.split('.').collect();
+    for part in parts {
+        let parsed_part = part.parse().with_context(|| "expected an integer")?;
+        parsed_parts.push(parsed_part);
+    }
+    let mut parsed = parsed_parts.into_iter();
+    Ok((
+        parsed.next().with_context(|| "missing number")?,
+        parsed.next().with_context(|| "missing number")?,
+        parsed.next().with_context(|| "missing number")?,
+    ))
+}
+
+pub(crate) fn version_as_array_or_wild(version: &str) -> anyhow::Result<BaseGameVersion> {
+    match version_as_array(version) {
+        Ok(v) => Ok(BaseGameVersion::Version(v.0, v.1, v.2)),
+        Err(_) if version == "*" => Ok(BaseGameVersion::Wild),
+        Err(_) => Err(anyhow!("expected `x.y.z` or `*`")),
+    }
 }