+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Allay adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Executables can now be used as an alternative to scripts (as described
+  [here](https://allay-mc.github.io/docs/scripts/executables.html)) (#15)
 - Add `schema` command (only available with `config-schema` feature).
 - Add `--bundle`, `--dir` and `--individual` flags for `build` command.
 - `--release` can now be invoked with `-r` as well.