+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Allay adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
-- Remove `development` field of `[project]` section.
+- Remove `development` field of `[project]` section (#21).
 - Remove `build_type` field of `[build]` section.