+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Allay adheres to [Semantic Versioning](https://semver.org/).
 
 ## [unreleased][]
 
+
+## [0.1.0][] - 2023-05-20
+
 _♻️ Rewrite_
 
 
@@ -35,6 +38,7 @@ _♻️ Rewrite_
 _🍰 Initial release_
 
 
-[unreleased]: https://github.com/allay-mc/allay/compare/v0.1.0-beta.1...HEAD
+[unreleased]: https://github.com/allay-mc/allay/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/allay-mc/allay/compare/v0.1.0-beta.1...v0.1.0
 [0.1.0-beta.1]: https://github.com/allay-mc/allay/compare/v0.1.0-alpha.1...v0.1.0-beta.1
 [0.1.0-alpha.1]: https://github.com/allay-mc/allay/releases/v0.1.0-alpha.1